Extract some types into their own files
